Refactor: organizing code in a more oo way
